packages/torrent_d/requirements.txt
lbry-libtorrent
requests
beautifulsoup4
lxml